app/utils/error_handling.py
# utils/error_handling.py
import logging
from tenacity import retry, stop_after_attempt, wait_exponential

logger = logging.getLogger(__name__)

class ErrorHandler:
    @retry(stop=stop_after_attempt(3), wait=wait_exponential(multiplier=1, min=4, max=10))
    def api_call_with_retry(self, func, *args, **kwargs):
        try:
            return func(*args, **kwargs)
        except Exception as e:
            logger.error(f"API call failed: {str(e)}")
            raise

class APIErrorHandler:
    @retry(stop=stop_after_attempt(3), wait=wait_exponential(multiplier=1, min=4, max=10))
    @handle_api_errors
    def safe_api_call(self, func, *args, **kwargs):
        return func(*args, **kwargs)